requirements.txt
streamlit==1.36.0
Pillow
requests==2.31.0
google-generativeai
python-dotenv==1.0.0
IPython